Add files via upload
In order to maximize efficiency/ reduce overhead of the single byte I/O operations, buffered byte streams were used for reading and writing both to and from the socket and input/output files (permitting control for customization of the buffer size). To avoid the deadlock, the parallel design approach was taken for reading and writing to/from the socket. Separate threads are used, were one thread (writeThread) is responsible for writing to the socket and another for reading (readThread) from the socket.
The two classes created to implement the Runnable interface are:
     1. ReadFromSocket
     2. WriteToSocket

The two classes are designed so that WriteToSocket handles the sending operations and ReadFromSocket handles the reading operations. Since the threads are implemented separately, .join() is used in order to:
     1. wait for writeToSocket thread to finish before closing the output stream after the file has transmitted
     2. close both the sockets once both threads are finished

In both the classes the custom buffer size and input/output streams are initialized in the class constructor. Additionally, files and socket are read/written to byte by byte (using an array of bytes to buffer the I/O) until they are signaled to terminate.